refactor(build-logic): streamline plugin and dependency management
Refactored several convention plugins to enhance consistency and reduce redundancy:
- Replaced hardcoded plugin and library applications with utility methods `getPlugin` and `getLibrary`.
- Simplified the Hilt, Room, and Test plugins by using `libs` for applying plugins and dependencies.
- Refined the `disableUnnecessaryAndroidTests` function and moved it under the `android` package for better organization.
- Improved dependency handling in the `AndroidKotlin` configuration using `implementation` and `ksp` extensions.
