+++ build-logic/convention/src/main/kotlin/AndroidLibraryConventionPlugin.kt
@@ -2,13 +2,14 @@ import com.android.build.api.variant.LibraryAndroidComponentsExtension
 import com.android.build.gradle.LibraryExtension
 import com.xbot.convention.Configuration
 import com.xbot.convention.android.configureAndroidKotlin
-import com.xbot.convention.disableUnnecessaryAndroidTests
+import com.xbot.convention.android.disableUnnecessaryAndroidTests
+import com.xbot.convention.extensions.getLibrary
 import com.xbot.convention.extensions.getPlugin
+import com.xbot.convention.extensions.implementation
 import com.xbot.convention.extensions.libs
 import org.gradle.api.Project
 import org.gradle.kotlin.dsl.configure
 import org.gradle.kotlin.dsl.dependencies
-import org.gradle.kotlin.dsl.kotlin
 
 class AndroidLibraryConventionPlugin : AndroidConventionPluginBase() {
 
@@ -29,10 +30,10 @@ class AndroidLibraryConventionPlugin : AndroidConventionPluginBase() {
         }
 
         dependencies {
-            add("androidTestImplementation", kotlin("test"))
-            add("testImplementation", kotlin("test"))
+//            androidTestImplementation(libs.getLibrary("kotlin-test"))
+//            testImplementation(libs.getLibrary("kotlin-test")
 
-            add("implementation", libs.findLibrary("androidx.tracing.ktx").get())
+            implementation(libs.getLibrary("androidx-tracing-ktx"))
         }
     }
 }
