+++ build-logic/convention/src/main/kotlin/AndroidTestConventionPlugin.kt
@@ -1,6 +1,8 @@
 import com.android.build.gradle.TestExtension
 import com.xbot.convention.Configuration
 import com.xbot.convention.android.configureAndroidKotlin
+import com.xbot.convention.extensions.getPlugin
+import com.xbot.convention.extensions.libs
 import org.gradle.api.Plugin
 import org.gradle.api.Project
 import org.gradle.kotlin.dsl.configure
@@ -9,8 +11,7 @@ class AndroidTestConventionPlugin : Plugin<Project> {
     override fun apply(target: Project) {
         with(target) {
             with(pluginManager) {
-                apply("com.android.test")
-                apply("org.jetbrains.kotlin.android")
+                apply(libs.getPlugin("android-test").get().pluginId)
             }
 
             extensions.configure<TestExtension> {
